adicionando validação dos campos de autor

+++ src/main/java/com/weberth/libraryapi/service/AutorService.java
@@ -1,8 +1,12 @@
 package com.weberth.libraryapi.service;
 
 import com.weberth.libraryapi.controller.dto.AutorDTO;
+import com.weberth.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.weberth.libraryapi.model.Autor;
 import com.weberth.libraryapi.repository.AutorRepository;
+import com.weberth.libraryapi.repository.LivroRepository;
+import com.weberth.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,16 +14,32 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository) {
-        this.repository = repository;
+//    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
+//        this.repository = repository;
+//        this.validator = validator;
+//        this.livroRepository = livroRepository;
+//    }
+
+    public Autor salvar(Autor autor)
+    {
+        validator.validar(autor);
+        return repository.save(autor);
     }
 
-    public Autor salvar(Autor autor){
-        return repository.save(autor);
+    public void atualizar(Autor autor){
+        if(autor.getId() == null){
+            throw new IllegalArgumentException("Para atualizar é necessário que o autor já esteja salvo na base.");
+        }
+        validator.validar(autor);
+
+        repository.save(autor);
     }
 
     public Optional<Autor> obterPorId(UUID id){
@@ -27,6 +47,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um autor que já possui livors cadastrados! ");
+        }
         repository.delete(autor);
     }
 
@@ -42,5 +65,10 @@ public class AutorService {
         }
         return repository.findAll();
     }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
+    }
+
 
 }
